BUGFIX: LOC-14
Added property betterRequestURI on WaybackReqeust, with a method to throw a BetterRequestException if it has been set.
ReplayServlet now calls the method to check if the request can be made with a diff URI, and implements the redirect if it is thrown.
ReplayFilter now checks if the URL requested is prefixed with the Wayback Context on which the request was made, and sets the betterRequestURI to the URL if this is the case.

git-svn-id: http://archive-access.svn.sourceforge.net/svnroot/archive-access/trunk/archive-access/projects@1491 69e27eb3-6e27-0410-b9c6-fffd7e226fab